gio: Remove deprecated use_boxed_functions override
`Gio.UnixMountPoint` likely still uses `use_boxed_functions` because
`g_unix_mount_point_copy` is only available on 2.54. However, the
`version` statement on the next line already ensures this type is only
available on that version or higher.

If anyone plans to support a lower version later on this can easily be
coded up in gir to replace the panic on version mismatch with some
`cfg_if` or `#[cfg()]` logic, just like what was done in [gstreamer-rs].

[gstreamer-rs]: https://gitlab.freedesktop.org/gstreamer/gstreamer-rs/-/merge_requests/677/diffs?commit_id=3c610e12e5a72f3d25949d84e5a4756b7545e34a