glib: Refactor glib::Value traits
The main change is that it for non-nullable types `Value::get()` will
never ever return `None`, replacing the previous confusing
`Value::get_some()`.

In addition it is now possible to `Value::get()` a nullable type
directly without `Option` wrapping. If the value was actually `None`
this would return an `Err`.

There is also a new generic `ValueType` trait that is implemented by all
types that can be used inside a `Value` and that allows
storing/retrieving types of this value into/from a `Value`.

Fixes https://github.com/gtk-rs/gtk-rs/issues/48